refactor: extract calculateTemperatureYAxisRanges to shared utility
- Create src/utils/chartUtils.ts with shared temperature buffer calculation
- Remove duplicate code from Field and Explorer temperature components
- Both components now import and use the same utility function
- Maintains same 0.5% buffer logic with DRY principle
